+++ pymc3/distributions/distribution.py
@@ -9,7 +9,7 @@ from ..model import (
     Model, get_named_nodes_and_relations, FreeRV,
     ObservedRV, MultiObservedRV, Context, InitContextMeta
 )
-from ..vartypes import string_types
+from ..vartypes import string_types, theano_constant
 from .shape_utils import (
     to_tuple,
     get_broadcastable_dist_samples,
@@ -92,7 +92,7 @@ class Distribution:
         if isinstance(val, tt.TensorVariable):
             return val.tag.test_value
 
-        if isinstance(val, tt.TensorConstant):
+        if isinstance(val, theano_constant):
             return val.value
 
         return val
@@ -238,9 +238,11 @@ class DensityDist(Distribution):
             The behavior of this callable can be altered with the
             ``wrap_random_with_dist_shape`` parameter.
             The supplied callable must have the following signature:
-            ``random(size=None, **kwargs)``, where ``size`` is the number of
-            IID draws to take from the distribution. Any extra keyword
-            argument can be added as required.
+            ``random(point=None, size=None, **kwargs)``, where ``point`` is a
+            ``None`` or a dictionary of random variable names and their
+            corresponding values (similar to what ``MultiTrace.get_point``
+            returns). ``size`` is the number of IID draws to take from the
+            distribution. Any extra keyword argument can be added as required.
         wrap_random_with_dist_shape: bool (Optional)
             If ``True``, the provided ``random`` callable is passed through
             ``generate_samples`` to make the random number generator aware of
@@ -502,7 +504,7 @@ class _DrawValuesContextBlocker(_DrawValuesContext, metaclass=InitContextMeta):
 def is_fast_drawable(var):
     return isinstance(var, (numbers.Number,
                             np.ndarray,
-                            tt.TensorConstant,
+                            theano_constant,
                             tt.sharedvar.SharedVariable))
 
 
@@ -592,7 +594,7 @@ def draw_values(params, point=None, size=None):
             if (next_, size) in drawn:
                 # If the node already has a givens value, skip it
                 continue
-            elif isinstance(next_, (tt.TensorConstant,
+            elif isinstance(next_, (theano_constant,
                                     tt.sharedvar.SharedVariable)):
                 # If the node is a theano.tensor.TensorConstant or a
                 # theano.tensor.sharedvar.SharedVariable, its value will be
@@ -783,7 +785,7 @@ def _draw_value(param, point=None, givens=None, size=None):
     """
     if isinstance(param, (numbers.Number, np.ndarray)):
         return param
-    elif isinstance(param, tt.TensorConstant):
+    elif isinstance(param, theano_constant):
         return param.value
     elif isinstance(param, tt.sharedvar.SharedVariable):
         return param.get_value()