+++ pymc3/exceptions.py
@@ -1,4 +1,9 @@
-__all__ = ['SamplingError', 'IncorrectArgumentsError', 'TraceDirectoryError']
+__all__ = [
+    "SamplingError",
+    "IncorrectArgumentsError",
+    "TraceDirectoryError",
+    "ImputationWarning",
+]
 
 
 class SamplingError(RuntimeError):
@@ -8,6 +13,14 @@ class SamplingError(RuntimeError):
 class IncorrectArgumentsError(ValueError):
     pass
 
+
 class TraceDirectoryError(ValueError):
-    '''Error from trying to load a trace from an incorrectly-structured directory,'''
+    """Error from trying to load a trace from an incorrectly-structured directory,"""
+
+    pass
+
+
+class ImputationWarning(UserWarning):
+    """Warning that there are missing values that will be imputed."""
+
     pass